old-tests/submission/W3C/domhtml/HTMLImageElement05.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IMG</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-img-element.html#dom-img-height" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<img alt="DTS IMAGE LOGO" height="50" src="/support/images/yellow-rectangle.png" width="100">
</p>



<script>
var domtest = async_test("The height attribute overrides the natural \"height\" of the image.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vheight;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("img");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vheight = testNode.height;

      assert_equals(vheight,50);
       
    domtest.done();
}));

</script>
<p>The height attribute overrides the natural "height" of the image.

    Retrieve the height attribute and examine it's value.</p>
<div id="log"></div>
</body></html>
